Removed query parameters handling from state
